Simplify precondition for RLRB recoloring
The LLRB adjustment created a right leaning red-black tree node. So it is
already known that the height of the tree didn't change during the
adjustment. But since the parent which was adjusted was handling an
misbalance in the right child tree, it is known that the tree under the
left tree must have one more black node. Thus the left child of the parent
must exist and must be black.