Simplify precondition for LLRB recoloring
Ff the current parent is for a misbalanced node from the right then there
must be a left node. Otherwise the LLRB would have been misbalanced before
the last erase because the removed node must have been a black node and
therefore it must also be at least one black (non NULL) node on the left
side. The same is true when the current parent is for a misbalanced node
from the left side.

The check for the right->left child can be omitted because the
!coming_from_right restructuring case already took care of it.