Avoid left rotate followed by right rotate on insert rebalance
A right leaning red node can either have a red or black sibling. The black
sibling has the benefit of a local fix by rotating the red node simply to
the left. This doesn't fix a possible misbalance but this has to be handled
by an upper layer in the tree anyway.

The red sibling on the other hand will be adjusted using a recolor (split
of 4-node into two 2-nodes + propagation of problem upwards the tree). But
the Sedgewick presentation from Maresias used a left rotate followed by a
right rotate before doing the actual recoloring. This is basically a null
operation which can be avoided.