Reduce rotates per adjustment for right subtree imbalance
The adjustment function for right subtrees with lower black height and red
node under right child of sibling can be implemented in different ways. The
initial implementation used 3 rotates and one recolor to fix the imbalance.

But it can also be implemented with 2 rotates and r recolors. But two of
the recolors can be done directly during the rotates. Thus the time for a
complete rotate can be saved.